Inject the Request on the ViewConfig
Injecting the request object on the view config adds the benefit
of no longer having to set the current language on the view config. This
prevents possible issues later in the lifecycle of this class.

The ViewConfigs scope had to be reduced to the Request scope. This
should pose no problems.