Allow token registration without email verification
Middleware can skip email verification when configured for an
institution. Selfservice now takes into account that configuration
option and sends the user directly to the last step of the
registration process when possession of the token is proven.

If email verification is enabled, there are four steps shown to the
user:

 1. Select token
 2. Link token (prove possession)
 3. Confirm (verify email)
 4. Activate token (vet the token)

If email verification is disabled, the third option is hidden.