Add 'print' and 'pdf' buttons on registration confirm page
The MPDF library is used to render the confirmation page
server-side. This output of the 'email sent' action is used to
generate the PDF. The print stylesheets are used by MPDF to get more
or less the same look and feel as the browser-printed webpage.

In order for the logo to be embedded in the PDF, php-gd is introduced
as a new dependency. If php-gd is not installed, a red cross will
appear on the PDF where the logo should be.