Collect Available tokens in a collection
When displaying the second factor tokens in the registration overview,
we want them to be sorted in LoA, alphabetical order. This was not
possible without sorting them in a collection as hard coded and gssp
tokens where passed individually to the template.

The new collection stores AvailableTokenInterface implementations. And
contains sorting logic.