Update button layout on token overview page
The test and add buttons are now shown in a seperate button group.
Previously the add and test buttons where placed in the table footer
and at the bottom of the page. This new solution allows us to group
the buttons next to each other, providing a cleaner page layout.