Override @font-face definition to prevent 404s
The Mopa Bootstrap bundle also allows us to load font awesome into the
frontend. This is done by including the bootstrap-fontawesome.less at
the top of the styles.less.

In turn the bs-fontawesome.less includes a path.less file that defines
the font-face variable. In our installed version, that font-face
definition results in 404's on the non existing woff2 font file.