Add a SmsSecondFactorService test double
This test double for now just assumes everything was alright and
accepts all verification & otp challenges.

The test double is put in place when SelfService runs in the `test`
environment.

A SmsSecondFactorServiceInterface was introduced in foresight of when we
start using dependency injection in the controllers.

Did some boyscouting during this commit (mainly cleanup of unused
imports).