Fix mpdf relative url's and logging
Relative url's weren't supported, because mpdf was trying to get it by
hostname in the request. If however a proxy should update the hostname
it wouldn't work. Therefore the `base` html tag was added which was
supported by mpdf for resolving absolute urls. Also the
logging was added and decent symfony shutdown.