Find verified second factors without authz context
The authorization context that was added for the RA application applies
an authorization context to the query. This filters out results that not
match the context of the current RA state (the institution that is
currently active).

For Self Service we do not want to apply that context based filtering.

See: https://www.pivotaltracker.com/story/show/165740432
See: https://github.com/OpenConext/Stepup-Middleware/pull/279
See: https://github.com/OpenConext/Stepup-Middleware-clientbundle/pull/82