Provide identity encryption functionality
The IdentityEncrypter is to be used to encrypt and write the attributes
received from one of the Remote vetting services (readid, idin or irma).
These attributes are encrypted with openssl public key encryption (can
be decrypted using the associated private key).

The important parts of this code is covered in test, an integration test
was written to prove the correctness of the encrypter.

No SAML response assertion is available yet from one of the services,
but a DTO is created that is to be used to contain these attributes. For
now bogus data is used to test the encrypter.

https://www.pivotaltracker.com/story/show/166932716