Make remote vetting IdPs further configurable
In the existing configuration yaml, a description and logo have been added. This allows for the better UI presentation of the RV IdP.

First version (home made and unverified) logos have been added for the IdPs

The config has been made available to the Twig templates via a view helper service. This service serves DTO objects that represent the IdP when displayed to the end user. So a title, description and logo can be retrieved from them.