Update the ecnryption logic in IdentityEncrypter
This implements aes-256-gcm symmetric encryption in the
IdentityEncrypter. This also changed the data that is saved to disk
slightly. Instead of only saving the encrypted data, other properties,
required for decrypting the data is also stored in the file.

The bulk of the work was provided by @pmeulen in:
https://gist.github.com/pmeulen/29da05c99165e3e4bfff8d7fbbb03fde

The code was integrated into the encrypter and updated to better fit the
code convetions of the project.

Tests have been updated according to the new setup. A decrypter was
added that aids the integration and unit test to verify the encryption
can successfully be decrypted.

This commit should satisfy requirements stated in this story:
https://www.pivotaltracker.com/story/show/171571420