Create AttributeCollectionAggregate
Used to aggregate AttributeCollectionInterface implementations. These are collected in the aggregate to be serialized to the file containing the encrypted IdP/RV data.

A composition of collection objects is achieved by adding the AttributeCollectionInterface onto the collection classes that are related to the aggregate. This way the aggregate can simply call the get attributes interface method on it's children and they will provide him with the required data that is safe to json encode.

One method changed on the AttributeListDto, the existing `getAttributes` was renamed to `getAttributeCollection`. The `getAttributes` method was also added, but that now provides an array representation of the collection.