Introduce the IdentityData object
This data object is filled with all the data the encrypter will be saving to disk. For now the object is instantiated in the constructor, but a factory will be introduced in a successive commit.

Other tasks that remain are adding the RV source. This is not yet easily possible at this point. But will also be implemented in an upcoming commit.

Other noticeable changes:

1. Introduced the ApplicationHelper, this helper creates the application version string (based on kernel root dir)
2. Moved the RemoteVettingService::fromAttributeSet method to the AttributeListDto class. As a static factory method. That way the RV service is not responsible for creating these kind of objects.
3. IdentityEncrypter now simply accepts a string as input.