Move IdentityAggregation to Remote vetting service
Move the aggregation of the IdentityData to the RemoteVettingService.
That way the controller will keep lightweight and the responsibility
of remote vetting is delegated and only glue is used in the
controller. This resulted in less coupling.

Also the provider name is set in the IdentityData, this needed to be
done after the feature was implemented on a different branch.