Self vetting (#227)
* Enable behat runs on SF4

* Import outlines of self vet logic based on RV

The RV (remote vetting) changes have been cherry picked from that
working branch. These changes have been rid of any RV references, but
the Self Vet option is implemented in exact manner. This should allow
for smooth merging later down the road.

* Update translations for self vetting

* Add a self vet marshaller

This service can determine if a user is allowed to self vet a token or not.
Based on institution configuration, and possession of a legitimate,
suitably equiped token.

* Update several config files

And install the new stepup middleware api bundle

* Integrate self vetting flow

Many changes are part of this commit. Some noteworthy changes are:

The move of all self vet controller logic to a dedicated controller. And
small fixes to previous work after excessive functional testing.

* Update translations and images

* Set licence headers to require SURFnet

This will probably be changed to SURF in the future, for now leave it at
the current requirement.

* Upgrade JavaScript dependencies

To prevent use of the vulnerable ellipse 3rd party dependency.

* Reword trans id's and translate flashbag messages

* Update index.php

The smoketest trigger is made to block use from prod environments

* Add remote-vetting css-class

* Update translations

* Show self-vetting first

* Update phpseclib

* Layout of vetting method selection

* update yarn packages

* update translations

* Update self-vet logo

* Evaluate if self_vet option enabled

The self vet option was not considered while determining if self vetting
is allowed. This was rectified in the self vet marshaller.

* Skip vetting type step if only one is available

* Rename remote-vetting-selector

Co-authored-by: Peter Havekes <peter.havekes@surf.nl>