Update EntryPointController [BSR]
The controller is now autowired. Dependency injection was
applied to get rid of accessing the service container directly.

In addition the service definition for the self asserted tokens
controller was taken down. As it is now able to autoload.

Finally, a logic change was applied, allowing display of the overview
action when recovery tokens are registered