Return false on 403 status in SelfAssertedTokensMarshaller
The exception when a 403 was returned was correctly interpreted as a not
allowed action. And that was not caught in the application. The
marshaller now tests for this.

This is probably a temporary construction as the authz endpoint will
start returning the http 200 status code on the authz requests that
resulted in a not allowed decision.