Express prefered token, and display this
When more than one token activation flows (vetting types) are available.
We show the preferred token activation mode by highlighting it with a
border color.

This solution is made without giving too much attention to presentation.
This because the new design will improve the activation flows
significantly. Having this feature in the current solution would be
costly. And we decided not to go through this at this point