Use Composer Audit instead of the local checker
Recent composer releases allow for auditing for secuirty
vulnerabilities. That is very convenient, and allows us to remove the
fabpot local security checker.

In addition the roave security checker is added to the project, that
also scans composer.lock for vulnerabilities during composer update and
dry runs. Adding an additional layer of security.