Set correct route for SMS resend button
During registration, the wrong route was visited for the sms resend
form. That resulted in the form being submitted unsuccessfully. By
testing if the secondfactor id is set (this is never the case when
adding a second RT from the token overview page) we can set the correct
route and params for both scenarios.