Use saml entitlement attribute for activation flow
Use the saml entitlement attribute to allow the activation flow.
If no activation could be allowed this will still result in service
desk vetting.

Also fixed the handling of imcoming query params which were not handled
correctly because the StepupSamlBundle handles the authentication
and the acivation flow param was therefore lost after entering first on
the consume-assertion endpoint in the application.

https://www.pivotaltracker.com/n/projects/1163646/stories/185558419